% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_links}
\alias{create_links}
\title{Create links between DELVE tables}
\usage{
create_links(data, parent, child)
}
\arguments{
\item{data}{Dataframe of DELVE table}

\item{parent}{Column containing IDs for parent table}

\item{child}{Column containing IDs (comma-separated strings) for child table}
}
\description{
Create links between DELVE tables
}
