% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_clip_loc}
\alias{extract_clip_loc}
\title{Extract marks from clip location field}
\usage{
extract_clip_loc(data)
}
\arguments{
\item{data}{Dataframe with clip_loc column}
}
\description{
Extract marks from clip location field
}
