% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_condition}
\alias{extract_condition}
\title{Extract condition from fish_condition field}
\usage{
extract_condition(data)
}
\arguments{
\item{data}{Dataframe with fish_condition column}
}
\description{
Extract condition from fish_condition field
}
