% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_mark}
\alias{extract_mark}
\title{Extract marks from mark status field}
\usage{
extract_mark(data)
}
\arguments{
\item{data}{Dataframe with mark_status column}
}
\description{
Extract marks from mark status field
}
