% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_other_mark_desc}
\alias{fix_other_mark_desc}
\title{Fix other mark description from DELVE tables}
\usage{
fix_other_mark_desc(x)
}
\arguments{
\item{x}{DELVE other mark description}
}
\description{
Fix other mark description from DELVE tables
}
