% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_velocity}
\alias{fix_velocity}
\title{Fix velocity values from DELVE tables}
\usage{
fix_velocity(vel, neg)
}
\arguments{
\item{x}{DELVE velocity value}
}
\description{
Fix velocity values from DELVE tables
}
