% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{from_canonical}
\alias{from_canonical}
\title{Convert canonical DELVE names to title case}
\usage{
from_canonical(x)
}
\arguments{
\item{x}{DELVE canonical name(s)}
}
\description{
Convert canonical DELVE names to title case
}
