% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone.R
\name{get_drone}
\alias{get_drone}
\title{Get Fulcrum drone tables}
\usage{
get_drone(table = c("main", "flight"))
}
\arguments{
\item{table}{Fulcrum table: main, flight}
}
\description{
Get Fulcrum drone tables
}
