% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_mark_rel.R
\name{get_emr}
\alias{get_emr}
\title{Get Fulcrum trap efficiency marking and release tables}
\usage{
get_emr(table = c("main", "marking", "release", "fl"))
}
\arguments{
\item{table}{Fulcrum table: main, marking, release, fl}
}
\description{
Get Fulcrum trap efficiency marking and release tables
}
