% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_transport.R
\name{get_ett}
\alias{get_ett}
\title{Get Fulcrum trap efficiency transport and tempering tables}
\usage{
get_ett(table = c("main", "transport", "transport_details", "tempering"))
}
\arguments{
\item{table}{Fulcrum table: main, transport, transport_details, tempering}
}
\description{
Get Fulcrum trap efficiency transport and tempering tables
}
