% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_processing.R
\name{get_fp}
\alias{get_fp}
\title{Get Fulcrum fish processing tables}
\usage{
get_fp(table = c("main", "fish", "plus"))
}
\arguments{
\item{table}{Fulcrum table: main, fish, plus}
}
\description{
Get Fulcrum fish processing tables
}
