% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_transport.R
\name{get_gct}
\alias{get_gct}
\title{Get Fulcrum trap general comments and tables}
\usage{
get_gct(table = c("main", "water_quality"))
}
\arguments{
\item{table}{Fulcrum table: main, water_quality}
}
\description{
Get Fulcrum trap general comments and tables
}
