% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_line.R
\name{get_hl}
\alias{get_hl}
\title{Get DELVE hook-and-line tables}
\usage{
get_hl(table = c("meta", "fish", "water_quality", "sensor"))
}
\arguments{
\item{table}{DELVE table: meta, fish, water_quality, sensor}
}
\description{
Get DELVE hook-and-line tables
}
