% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{get_si}
\alias{get_si}
\title{Get DELVE structure inspection tables}
\usage{
get_si(
  table = c("meta", "change", "boom", "guidance", "curtain", "wildlife", "listed",
    "anchor", "dock", "trap", "velocity", "float")
)
}
\arguments{
\item{table}{DELVE table: meta, change, boom, guidance, curtain, wildlife, listed, anchor, dock, trap, velocity, float}
}
\description{
Get DELVE structure inspection tables
}
