% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomach_sampling.R
\name{get_ss}
\alias{get_ss}
\title{Get Fulcrum stomach sampling tables}
\usage{
get_ss(table = c("main", "predator", "contents"))
}
\arguments{
\item{table}{Fulcrum table: main, predator, contents}
}
\description{
Get Fulcrum stomach sampling tables
}
