% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{get_ssi}
\alias{get_ssi}
\title{Get Fulcrum structure and safety inspection tables}
\usage{
get_ssi(
  table = c("main", "debris_boom", "net_anchor", "net_dock", "structure")
)
}
\arguments{
\item{table}{Fulcrum table: main, debris_boom, net_anchor, net_dock, structure}
}
\description{
Get Fulcrum structure and safety inspection tables
}
