% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_buoy.R
\name{get_tb}
\alias{get_tb}
\title{Get Fulcrum temperature buoy tables}
\usage{
get_tb(table = c("main", "buoy", "sensor"))
}
\arguments{
\item{table}{Fulcrum table: main, buoy, sensor}
}
\description{
Get Fulcrum temperature buoy tables
}
