% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{get_tc}
\alias{get_tc}
\title{Get DELVE trap collection tables}
\usage{
get_tc(
  table = c("meta", "fish", "plus", "stomach", "transport", "transport_wq")
)
}
\arguments{
\item{table}{DELVE table: meta, fish, tally, stomach, transport, transport_wq}
}
\description{
Get DELVE trap collection tables
}
