% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{get_te}
\alias{get_te}
\title{Get DELVE trap efficiency tables}
\usage{
get_te(
  table = c("meta", "fish", "transport", "transport_wq", "tempering", "holding",
    "release")
)
}
\arguments{
\item{table}{DELVE table: meta, fish, transport, transport_wq, tempering, holding, release}
}
\description{
Get DELVE trap efficiency tables
}
