% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_platform.R
\name{get_tp}
\alias{get_tp}
\title{Get Fulcrum trap platform tables}
\usage{
get_tp(table = c("main", "perf_config", "velocity", "water_quality"))
}
\arguments{
\item{table}{Fulcrum table: main, perf_config, velocity, water_quality}
}
\description{
Get Fulcrum trap platform tables
}
