% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{get_wq}
\alias{get_wq}
\title{Get DELVE water quality tables}
\usage{
get_wq(
  table = c("meta", "station", "retrieval", "deployment", "equipment", "calibration")
)
}
\arguments{
\item{table}{DELVE table: meta, station, retrieval, deployment, equipment, calibration,}
}
\description{
Get DELVE water quality tables
}
