% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi.R
\name{get_ysi}
\alias{get_ysi}
\title{Get Fulcrum YSI tables}
\usage{
get_ysi(table = c("main", "sample", "depth"))
}
\arguments{
\item{table}{Fulcrum table: main, sample, depth}
}
\description{
Get Fulcrum YSI tables
}
