% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_datetime}
\alias{make_datetime}
\title{Make Fulcrum datetime from date and time fields}
\usage{
make_datetime(date, time, tz = tz_loc)
}
\arguments{
\item{date}{Fulcrum date field}

\item{time}{Fulcrum time field}
}
\description{
Make Fulcrum datetime from date and time fields
}
