% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{prep_adcp_main}
\alias{prep_adcp_main}
\title{Prepare Fulcrum ADCP main table}
\usage{
prep_adcp_main(adcp_main_raw)
}
\arguments{
\item{adcp_main_raw}{Unprocessed ADCP main table}
}
\description{
Prepare Fulcrum ADCP main table
}
