% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{prep_adcp_transect}
\alias{prep_adcp_transect}
\title{Prepare Fulcrum ADCP transect table}
\usage{
prep_adcp_transect(adcp_transect_raw, adcp_main)
}
\arguments{
\item{adcp_transect_raw}{Unprocessed ADCP transect table}

\item{adcp_main}{Processed ADCP main table}
}
\description{
Prepare Fulcrum ADCP transect table
}
