% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone.R
\name{prep_drone_flight}
\alias{prep_drone_flight}
\title{Prepare Fulcrum drone flight table}
\usage{
prep_drone_flight(drone_flight_raw, drone_main)
}
\arguments{
\item{drone_flight_raw}{Unprocessed drone flight table}

\item{drone_main}{Processed drone main table}
}
\description{
Prepare Fulcrum drone flight table
}
