% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drone.R
\name{prep_drone_main}
\alias{prep_drone_main}
\title{Prepare Fulcrum drone main table}
\usage{
prep_drone_main(drone_main_raw)
}
\arguments{
\item{drone_main_raw}{Unprocessed drone main table}
}
\description{
Prepare Fulcrum drone main table
}
