% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_dt_fulcrum}
\alias{prep_dt_fulcrum}
\title{Make Fulcrum datetime from date and time fields}
\usage{
prep_dt_fulcrum(date, time, tz = tz_loc)
}
\arguments{
\item{date}{Fulcrum date field}

\item{time}{Fulcrum time field}

\item{tz}{Time zone}
}
\description{
Make Fulcrum datetime from date and time fields
}
