% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_mark_rel.R
\name{prep_emr_fl}
\alias{prep_emr_fl}
\title{Prepare Fulcrum efficiency marking fork length table}
\usage{
prep_emr_fl(emr_fl_raw, emr_mark)
}
\arguments{
\item{emr_fl_raw}{Unprocessed efficiency marking fork length table}

\item{emr_mark}{Processed marking table}
}
\description{
Prepare Fulcrum efficiency marking fork length table
}
