% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_mark_rel.R
\name{prep_emr_main}
\alias{prep_emr_main}
\title{Prepare Fulcrum efficiency marking and release main table}
\usage{
prep_emr_main(emr_main_raw)
}
\arguments{
\item{emr_main_raw}{Unprocessed marking and release main table}
}
\description{
Prepare Fulcrum efficiency marking and release main table
}
