% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_mark_rel.R
\name{prep_emr_mark}
\alias{prep_emr_mark}
\title{Prepare Fulcrum efficiency marking table}
\usage{
prep_emr_mark(emr_mark_raw, emr_main)
}
\arguments{
\item{emr_mark_raw}{Unprocessed efficiency marking table}

\item{emr_main}{Processed marking and release main table}
}
\description{
Prepare Fulcrum efficiency marking table
}
