% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_mark_rel.R
\name{prep_emr_rel}
\alias{prep_emr_rel}
\title{Prepare Fulcrum efficiency release table}
\usage{
prep_emr_rel(emr_rel_raw, emr_main)
}
\arguments{
\item{emr_rel_raw}{Unprocessed efficiency release table}

\item{emr_main}{Processed marking and release main table}
}
\description{
Prepare Fulcrum efficiency release table
}
