% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_transport.R
\name{prep_ett_main}
\alias{prep_ett_main}
\title{Prepare Fulcrum efficiency transport and tempering main table}
\usage{
prep_ett_main(ett_main_raw)
}
\arguments{
\item{ett_main_raw}{Unprocessed transport and tempering main table}
}
\description{
Prepare Fulcrum efficiency transport and tempering main table
}
