% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_transport.R
\name{prep_ett_tempering}
\alias{prep_ett_tempering}
\title{Prepare Fulcrum efficiency tempering table}
\usage{
prep_ett_tempering(ett_tempering_raw, ett_main)
}
\arguments{
\item{ett_tempering_raw}{Unprocessed efficiency tempering table}

\item{ett_main}{Processed transport and tempering main table}
}
\description{
Prepare Fulcrum efficiency tempering table
}
