% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_transport.R
\name{prep_ett_transport}
\alias{prep_ett_transport}
\title{Prepare Fulcrum efficiency transport table}
\usage{
prep_ett_transport(ett_transport_raw, ett_td_raw, ett_main)
}
\arguments{
\item{ett_transport_raw}{Unprocessed efficiency transport table}

\item{ett_td_raw}{Unprocessed efficiency transport details table}

\item{ett_main}{Processed transport and tempering main table}
}
\description{
Prepare Fulcrum efficiency transport table
}
