% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_processing.R
\name{prep_fp_fish}
\alias{prep_fp_fish}
\title{Prepare Fulcrum fish measurements table}
\usage{
prep_fp_fish(fp_fish_raw, fp_main)
}
\arguments{
\item{fp_fish_raw}{Unprocessed fish measurements table}

\item{fp_main}{Processed fish processing main table}
}
\description{
Prepare Fulcrum fish measurements table
}
