% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_processing.R
\name{prep_fp_main}
\alias{prep_fp_main}
\title{Prepare Fulcrum fish processing main table}
\usage{
prep_fp_main(fp_main_raw)
}
\arguments{
\item{fp_main_raw}{Unprocessed fish processing main table}
}
\description{
Prepare Fulcrum fish processing main table
}
