% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_processing.R
\name{prep_fp_plus}
\alias{prep_fp_plus}
\title{Prepare Fulcrum plus counts table}
\usage{
prep_fp_plus(fp_plus_raw, fp_main)
}
\arguments{
\item{fp_plus_raw}{Unprocessed plus counts table}

\item{fp_main}{Processed fish processing main table}
}
\description{
Prepare Fulcrum plus counts table
}
