% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_transport.R
\name{prep_gct_gc}
\alias{prep_gct_gc}
\title{Prepare Fulcrum general comments table}
\usage{
prep_gct_gc(gct_main)
}
\arguments{
\item{gct_main}{Processed comments and transport main table}
}
\description{
Prepare Fulcrum general comments table
}
