% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_transport.R
\name{prep_gct_main}
\alias{prep_gct_main}
\title{Prepare Fulcrum general comments and transport main table}
\usage{
prep_gct_main(gct_main_raw)
}
\arguments{
\item{gct_main_raw}{Unprocessed comments and transport main table}
}
\description{
Prepare Fulcrum general comments and transport main table
}
