% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_transport.R
\name{prep_gct_transport}
\alias{prep_gct_transport}
\title{Prepare Fulcrum transport main table}
\usage{
prep_gct_transport(gct_main)
}
\arguments{
\item{gct_main}{Processed comments and transport main table}
}
\description{
Prepare Fulcrum transport main table
}
