% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_transport.R
\name{prep_gct_transport_wq}
\alias{prep_gct_transport_wq}
\title{Prepare Fulcrum transport water quality table}
\usage{
prep_gct_transport_wq(gct_wq_raw, gct_transport)
}
\arguments{
\item{gct_wq_raw}{Unprocessed transport water quality table}

\item{gct_main}{Processed transport  main table}
}
\description{
Prepare Fulcrum transport water quality table
}
