% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_line.R
\name{prep_hl_meta}
\alias{prep_hl_meta}
\title{Prepare DELVE hook-and-line metadata}
\usage{
prep_hl_meta(hl_meta_raw, min_date = min_hl_date, tz = tz_loc)
}
\arguments{
\item{hl_meta_raw}{Unprocessed hook-and-line metadata}

\item{min_date}{First date to use for hook-and-line data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE hook-and-line metadata
}
