% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_line.R
\name{prep_hl_sensor}
\alias{prep_hl_sensor}
\title{Prepare DELVE hook-and-line water quality sensor data}
\usage{
prep_hl_sensor(hl_sensor_raw, hl_wq_raw, hl_wq)
}
\arguments{
\item{hl_sensor_raw}{Unprocessed hook-and-line water quality sensor data}

\item{hl_wq_raw}{Unprocessed hook-and-line water quality data}

\item{hl_wq}{Processed hook-and-line water quality data}
}
\description{
Prepare DELVE hook-and-line water quality sensor data
}
