% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_anchor}
\alias{prep_si_anchor}
\title{Prepare DELVE anchor inspection data}
\usage{
prep_si_anchor(si_anchor_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{si_anchor_raw}{Unprocessed anchor inspection data}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE anchor inspection data
}
