% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_boom}
\alias{prep_si_boom}
\title{Prepare DELVE boom inspection data}
\usage{
prep_si_boom(si_boom_raw, si_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{si_boom_raw}{Unprocessed boom inspection data}

\item{si_meta_raw}{Unprocessed structure inspection metadata}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE boom inspection data
}
