% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_change}
\alias{prep_si_change}
\title{Prepare DELVE structure inspection change log data}
\usage{
prep_si_change(si_change_raw, si_meta)
}
\arguments{
\item{si_change_raw}{Unprocessed structure inspection change log data}

\item{si_meta}{Processed structure inspection metadata}
}
\description{
Prepare DELVE structure inspection change log data
}
