% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_curtain}
\alias{prep_si_curtain}
\title{Prepare DELVE temperature curtain inspection data}
\usage{
prep_si_curtain(
  si_curtain_raw,
  si_meta_raw,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{si_curtain_raw}{Unprocessed temperature curtain inspection data}

\item{si_meta_raw}{Unprocessed structure inspection metadata}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE temperature curtain inspection data
}
