% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_float}
\alias{prep_si_float}
\title{Prepare DELVE float and curtain inspection data}
\usage{
prep_si_float(
  si_float_raw,
  si_curtain_raw,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{si_float_raw}{Unprocessed float and curtain inspection data}

\item{si_curtain_raw}{Unprocessed temperature curtain inspection data}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE float and curtain inspection data
}
