% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_listed}
\alias{prep_si_listed}
\title{Prepare DELVE listed species impacts}
\usage{
prep_si_listed(si_listed_raw, si_wildlife_raw, si_wildlife)
}
\arguments{
\item{si_listed_raw}{Unprocessed listed species impacts}

\item{si_wildlife_raw}{Unprocessed wildlife observations}

\item{si_wildlife}{Processed wildlife observations}
}
\description{
Prepare DELVE listed species impacts
}
