% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_meta}
\alias{prep_si_meta}
\title{Prepare DELVE structure inspection metadata}
\usage{
prep_si_meta(si_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{si_meta_raw}{Unprocessed structure inspection metadata}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE structure inspection metadata
}
