% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_trap}
\alias{prep_si_trap}
\title{Prepare DELVE trap platform data}
\usage{
prep_si_trap(si_trap_raw, si_guidance_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{si_trap_raw}{Unprocessed trap platform data}

\item{si_guidance_raw}{Unprocessed guidance net inspection data}

\item{min_date}{First date to use for structure inspection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap platform data
}
