% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_velocity}
\alias{prep_si_velocity}
\title{Prepare DELVE trap velocity data}
\usage{
prep_si_velocity(si_velocity_raw, si_trap)
}
\arguments{
\item{si_velocity_raw}{Unprocessed trap velocity data}

\item{si_trap}{Processed trap platform data}
}
\description{
Prepare DELVE trap velocity data
}
