% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_inspection.R
\name{prep_si_wildlife}
\alias{prep_si_wildlife}
\title{Prepare DELVE wildlife observations}
\usage{
prep_si_wildlife(si_wildlife_raw, si_meta_raw, si_meta)
}
\arguments{
\item{si_wildlife_raw}{Unprocessed wildlife observations}

\item{si_meta_raw}{Unprocessed structure inspection metadata}

\item{si_meta}{Processed structure inspection metadata}
}
\description{
Prepare DELVE wildlife observations
}
