% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomach_sampling.R
\name{prep_ss_contents}
\alias{prep_ss_contents}
\title{Prepare Fulcrum stomach contents table}
\usage{
prep_ss_contents(ss_contents_raw, ss_main)
}
\arguments{
\item{ss_contents_raw}{Unprocessed stomach contents table}

\item{ss_main}{Processed stomach sammpling main table}
}
\description{
Prepare Fulcrum stomach contents table
}
