% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomach_sampling.R
\name{prep_ss_main}
\alias{prep_ss_main}
\title{Prepare Fulcrum stomach sampling main table}
\usage{
prep_ss_main(ss_main_raw)
}
\arguments{
\item{ss_main_raw}{Unprocessed stomach sampling main table}
}
\description{
Prepare Fulcrum stomach sampling main table
}
