% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomach_sampling.R
\name{prep_ss_predator}
\alias{prep_ss_predator}
\title{Prepare Fulcrum predator table}
\usage{
prep_ss_predator(ss_predator_raw, ss_main)
}
\arguments{
\item{ss_predator_raw}{Unprocessed predator table}

\item{ss_main}{Processed stomach sampling main table}
}
\description{
Prepare Fulcrum predator table
}
