% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{prep_ssi_debris_boom}
\alias{prep_ssi_debris_boom}
\title{Prepare Fulcrum debris boom inspection table}
\usage{
prep_ssi_debris_boom(ssi_debris_boom_raw, ssi_main)
}
\arguments{
\item{ssi_debris_boom_raw}{Unprocessed debris boom inspection table}

\item{ssi_main}{Processed structure and safety main table}
}
\description{
Prepare Fulcrum debris boom inspection table
}
