% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{prep_ssi_main}
\alias{prep_ssi_main}
\title{Prepare Fulcrum structure and safety main table}
\usage{
prep_ssi_main(ssi_main_raw)
}
\arguments{
\item{ssi_main_raw}{Unprocessed structure and safety main table}
}
\description{
Prepare Fulcrum structure and safety main table
}
