% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{prep_ssi_net_anchor}
\alias{prep_ssi_net_anchor}
\title{Prepare Fulcrum guidance net anchor inspection table}
\usage{
prep_ssi_net_anchor(ssi_net_anchor_raw, ssi_main)
}
\arguments{
\item{ssi_net_anchor_raw}{Unprocessed guidance net anchor inspection table}

\item{ssi_main}{Processed structure and safety main table}
}
\description{
Prepare Fulcrum guidance net anchor inspection table
}
