% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{prep_ssi_net_dock}
\alias{prep_ssi_net_dock}
\title{Prepare Fulcrum guidance net dock inspection table}
\usage{
prep_ssi_net_dock(ssi_net_dock_raw, ssi_main)
}
\arguments{
\item{ssi_net_dock_raw}{Unprocessed guidance net dock inspection table}

\item{ssi_main}{Processed structure and safety main table}
}
\description{
Prepare Fulcrum guidance net dock inspection table
}
