% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_safety.R
\name{prep_ssi_structure}
\alias{prep_ssi_structure}
\title{Prepare Fulcrum structure inspection table}
\usage{
prep_ssi_structure(ssi_structure_raw, ssi_main)
}
\arguments{
\item{ssi_structure_raw}{Unprocessed guidance structure inspection table}

\item{ssi_main}{Processed structure and safety main table}
}
\description{
Prepare Fulcrum structure inspection table
}
