% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_buoy.R
\name{prep_tb_buoy}
\alias{prep_tb_buoy}
\title{Prepare Fulcrum temperature buoy table}
\usage{
prep_tb_buoy(tb_buoy_raw, tb_main)
}
\arguments{
\item{tb_buoy_raw}{Unprocessed temperature buoy table}

\item{tb_main}{Processed temperature buoy main table}
}
\description{
Prepare Fulcrum temperature buoy table
}
