% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_buoy.R
\name{prep_tb_main}
\alias{prep_tb_main}
\title{Prepare Fulcrum temperature buoy main table}
\usage{
prep_tb_main(tb_main_raw)
}
\arguments{
\item{tb_main_raw}{Unprocessed temperature buoy main table}
}
\description{
Prepare Fulcrum temperature buoy main table
}
