% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_buoy.R
\name{prep_tb_sensor}
\alias{prep_tb_sensor}
\title{Prepare Fulcrum temperature sensor table}
\usage{
prep_tb_sensor(tb_sensor_raw, tb_buoy)
}
\arguments{
\item{tb_sensor_raw}{Unprocessed temperature sensor table}

\item{tb_buoy}{Processed temperature buoy table}
}
\description{
Prepare Fulcrum temperature sensor table
}
