% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap.R
\name{prep_tc_fish}
\alias{prep_tc_fish}
\title{Prepare DELVE trap collection fish measurement data}
\usage{
prep_tc_fish(
  tc_fish_raw,
  tc_meta_raw,
  tc_meta,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{tc_meta_raw}{Unprocessed trap collection metadata}

\item{tc_meta}{Processed trap collection metadata}

\item{min_date}{First date to use for trap collection data}

\item{tz}{Time zone name}

\item{tc_meta_fish}{Unprocessed trap collection fish measurement data}
}
\description{
Prepare DELVE trap collection fish measurement data
}
