% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_fish_stomach}
\alias{prep_tc_fish_stomach}
\title{Prepare DELVE trap collection stomach data from the fish measurement table}
\usage{
prep_tc_fish_stomach(tc_fish_raw, tc_meta_raw, tz = tz_loc)
}
\arguments{
\item{tc_fish_raw}{Unprocessed trap collection fish measurement data}

\item{tc_meta_raw}{Unprocessed trap collection metadata}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap collection stomach data from the fish measurement table
}
