% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_meta}
\alias{prep_tc_meta}
\title{Prepare DELVE trap collection metadata}
\usage{
prep_tc_meta(tc_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{tc_meta_raw}{Unprocessed trap collection metadata}

\item{min_date}{First date to use for trap collection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap collection metadata
}
