% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_plus}
\alias{prep_tc_plus}
\title{Prepare DELVE trap collection plus count data}
\usage{
prep_tc_plus(tc_plus_raw, tc_meta_raw, tc_meta, min_date = min_tc_date)
}
\arguments{
\item{tc_plus_raw}{Unprocessed trap collection plus count data}

\item{tc_meta_raw}{Unprocessed trap collection metadata}

\item{tc_meta}{Processed trap collection metadata}

\item{min_date}{First date to use for trap collection data}
}
\description{
Prepare DELVE trap collection plus count data
}
