% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_plus_stomach}
\alias{prep_tc_plus_stomach}
\title{Prepare DELVE trap collection stomach data from the plus count table}
\usage{
prep_tc_plus_stomach(tc_plus_raw, tc_meta_raw, tc_meta)
}
\arguments{
\item{tc_plus_raw}{Unprocessed trap collection plus count data}

\item{tc_meta_raw}{Unprocessed trap collection metadata}

\item{tc_meta}{Processed trap collection metadata}
}
\description{
Prepare DELVE trap collection stomach data from the plus count table
}
