% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_stomach}
\alias{prep_tc_stomach}
\title{Prepare DELVE trap collection stomach data}
\usage{
prep_tc_stomach(tc_stomach_raw, tc_fish_raw, tc_fish, tc_plus)
}
\arguments{
\item{tc_stomach_raw}{Unprocessed trap collection stomach data}

\item{tc_fish_raw}{Unprocessed trap collection fish measurement data}

\item{tc_fish}{Processed trap collection fish measurement data}

\item{tc_plus}{Processed trap collection plus count data}

\item{min_date}{First date to use for trap collection data}
}
\description{
Prepare DELVE trap collection stomach data
}
