% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_transport_wq}
\alias{prep_tc_transport_wq}
\title{Prepare DELVE trap collection transport water quality data}
\usage{
prep_tc_transport_wq(
  tc_transport_wq_raw,
  tc_transport_raw,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{tc_transport_wq_raw}{Unprocessed trap collection transport data}

\item{tc_transport_raw}{Unprocessed trap collection transport data}

\item{min_date}{First date to use for trap collection data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap collection transport water quality data
}
