% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{prep_te_fish}
\alias{prep_te_fish}
\title{Prepare DELVE trap efficiency fish data}
\usage{
prep_te_fish(te_fish_raw, tc_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{te_fish_raw}{Unprocessed trap collection fish measurement data}

\item{min_date}{First date to use for trap efficiency data}

\item{tz}{Time zone name}

\item{te_meta_raw}{Unprocessed trap collection metadata}
}
\description{
Prepare DELVE trap efficiency fish data
}
