% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{prep_te_holding}
\alias{prep_te_holding}
\title{Prepare DELVE trap efficiency transport holding data}
\usage{
prep_te_holding(te_holding_raw, te_transport_raw, te_transport)
}
\arguments{
\item{te_holding_raw}{Unprocessed trap efficiency transport holding data}

\item{te_transport_raw}{Unprocessed trap efficiency transport data}

\item{te_transport}{Processed trap efficiency transport data}
}
\description{
Prepare DELVE trap efficiency transport holding data
}
