% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{prep_te_meta}
\alias{prep_te_meta}
\title{Prepare DELVE trap efficiency metadata}
\usage{
prep_te_meta(te_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{te_meta_raw}{Unprocessed trap collection metadata}

\item{min_date}{First date to use for trap efficiency data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap efficiency metadata
}
