% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{prep_te_tempering}
\alias{prep_te_tempering}
\title{Prepare DELVE trap efficiency transport tempering data}
\usage{
prep_te_tempering(te_tempering_raw, te_transport_raw, te_transport)
}
\arguments{
\item{te_tempering_raw}{Unprocessed trap efficiency transport tempering data}

\item{te_transport_raw}{Unprocessed trap efficiency transport data}

\item{te_transport}{Processed trap efficiency transport data}
}
\description{
Prepare DELVE trap efficiency transport tempering data
}
