% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_efficiency.R
\name{prep_te_transport}
\alias{prep_te_transport}
\title{Prepare DELVE trap efficiency transport data}
\usage{
prep_te_transport(te_transport_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{te_transport_raw}{Unprocessed trap efficiency transport data}

\item{min_date}{First date to use for trap efficiency data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE trap efficiency transport data
}
