% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_time_fulcrum}
\alias{prep_time_fulcrum}
\title{Make time string from Fulcrum time field}
\usage{
prep_time_fulcrum(time)
}
\arguments{
\item{time}{Fulcrum time field}
}
\description{
Make time string from Fulcrum time field
}
