% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_platform.R
\name{prep_tp_main}
\alias{prep_tp_main}
\title{Prepare Fulcrum trap platform main table}
\usage{
prep_tp_main(tp_main_raw)
}
\arguments{
\item{tp_main_raw}{Unprocessed trap platform main table}
}
\description{
Prepare Fulcrum trap platform main table
}
