% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_platform.R
\name{prep_tp_perf_config}
\alias{prep_tp_perf_config}
\title{Prepare Fulcrum performance and configuration table}
\usage{
prep_tp_perf_config(tp_perf_config_raw, tp_main)
}
\arguments{
\item{tp_perf_config_raw}{Unprocessed performance and configuration table}

\item{tp_main}{Processed trap platform main table}
}
\description{
Prepare Fulcrum performance and configuration table
}
