% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_platform.R
\name{prep_tp_velocity}
\alias{prep_tp_velocity}
\title{Prepare Fulcrum velocity table}
\usage{
prep_tp_velocity(tp_velocity_raw, tp_main)
}
\arguments{
\item{tp_velocity_raw}{Unprocessed velocity table}

\item{tp_main}{Processed fish processing main table}
}
\description{
Prepare Fulcrum velocity table
}
