% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_platform.R
\name{prep_tp_wq}
\alias{prep_tp_wq}
\title{Prepare Fulcrum velocity table}
\usage{
prep_tp_wq(tp_wq_raw, tp_main)
}
\arguments{
\item{tp_wq_raw}{Unprocessed water quality table}

\item{tp_main}{Processed trap platform main table}
}
\description{
Prepare Fulcrum velocity table
}
