% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{prep_wq_deployment}
\alias{prep_wq_deployment}
\title{Prepare DELVE water quality deployment data}
\usage{
prep_wq_deployment(
  wq_deployment_raw,
  wq_station_raw,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{wq_deployment_raw}{Unprocessed water quality deployment data}

\item{wq_station_raw}{Unprocessed water quality station metadata}

\item{min_date}{First date to use for water quality data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE water quality deployment data
}
