% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{prep_wq_equipment}
\alias{prep_wq_equipment}
\title{Prepare DELVE water quality equipment data}
\usage{
prep_wq_equipment(wq_equipment_raw, wq_retrieval_raw, wq_deployment_raw)
}
\arguments{
\item{wq_equipment_raw}{Unprocessed water quality equipment data}

\item{wq_retrieval_raw}{Unprocessed water quality retrieval data}

\item{wq_deployment_raw}{Unprocessed water quality deployment data}
}
\description{
Prepare DELVE water quality equipment data
}
