% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{prep_wq_meta}
\alias{prep_wq_meta}
\title{Prepare DELVE water quality metadata}
\usage{
prep_wq_meta(wq_meta_raw, min_date = min_tc_date, tz = tz_loc)
}
\arguments{
\item{wq_meta_raw}{Unprocessed water quality metadata}

\item{min_date}{First date to use for water quality data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE water quality metadata
}
