% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{prep_wq_retrieval}
\alias{prep_wq_retrieval}
\title{Prepare DELVE water quality retrieval data}
\usage{
prep_wq_retrieval(
  wq_retrieval_raw,
  wq_station_raw,
  min_date = min_tc_date,
  tz = tz_loc
)
}
\arguments{
\item{wq_retrieval_raw}{Unprocessed water quality retrieval data}

\item{wq_station_raw}{Unprocessed water quality station metadata}

\item{min_date}{First date to use for water quality data}

\item{tz}{Time zone name}
}
\description{
Prepare DELVE water quality retrieval data
}
