% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_quality.R
\name{prep_wq_station}
\alias{prep_wq_station}
\title{Prepare DELVE water quality station data}
\usage{
prep_wq_station(wq_station_raw, wq_meta_raw, wq_meta)
}
\arguments{
\item{wq_station_raw}{Unprocessed water quality station data}

\item{wq_meta_raw}{Unprocessed water quality metadata}

\item{wq_meta}{Processed water quality metadata}
}
\description{
Prepare DELVE water quality station data
}
