% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi.R
\name{prep_ysi_depth}
\alias{prep_ysi_depth}
\title{Prepare Fulcrum YSI depth table}
\usage{
prep_ysi_depth(ysi_depth_raw, ysi_sample)
}
\arguments{
\item{ysi_depth_raw}{Unprocessed YSI depth table}

\item{ysi_sample}{Processed YSI sample table}
}
\description{
Prepare Fulcrum YSI depth table
}
