% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi.R
\name{prep_ysi_main}
\alias{prep_ysi_main}
\title{Prepare Fulcrum YSI main table}
\usage{
prep_ysi_main(ysi_main_raw)
}
\arguments{
\item{ysi_main_raw}{Unprocessed YSI main table}
}
\description{
Prepare Fulcrum YSI main table
}
