% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysi.R
\name{prep_ysi_sample}
\alias{prep_ysi_sample}
\title{Prepare Fulcrum YSI sample table}
\usage{
prep_ysi_sample(ysi_sample_raw, ysi_main)
}
\arguments{
\item{ysi_sample_raw}{Unprocessed YSI sample table}

\item{ysi_main}{Processed YSI main table}
}
\description{
Prepare Fulcrum YSI sample table
}
