% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap.R
\name{prep_tc_meta}
\alias{prep_tc_meta}
\title{Prepare DELVE trap collection metadata}
\usage{
prep_tc_meta(tc_meta_raw)
}
\arguments{
\item{tc_meta_raw}{Unprocessed trap collection metadata}
}
\description{
Prepare DELVE trap collection metadata
}
