% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_collection.R
\name{prep_tc_stomach}
\alias{prep_tc_stomach}
\title{Prepare DELVE trap collection stomach data}
\usage{
prep_tc_stomach(
  tc_stomach_raw,
  tc_fish_raw,
  tc_fish,
  tc_fish_stomach,
  tc_plus_stomach
)
}
\arguments{
\item{tc_stomach_raw}{Unprocessed trap collection stomach data}

\item{tc_fish_raw}{Unprocessed trap collection fish measurement data}

\item{tc_fish}{Processed trap collection fish measurement data}

\item{tc_fish_stomach}{Processed trap collection stomach data from fish measurement table}

\item{tc_plus_stomach}{Processed trap collection stomach data from plus count table}
}
\description{
Prepare DELVE trap collection stomach data
}
